Fix & Update & Optimization

+++ backend/controllers/admin/admin.user.ts
@@ -8,7 +8,7 @@ import Inventory from '../../models/shop/inventory.model';
 import Report from '../../models/report.model';
 import Address from '../../models/shop/address.model';
 import Comment from '../../models/shop/comment.model';
-import type { ICart, IComment, IPagination, IProduct, IReport, IUser, IWishList } from '../../types';
+import type { IPagination, IProduct, IUser } from '../../types';
 
 export const users = async (req : Request, res : Response, next : NextFunction) => {
 
@@ -25,7 +25,7 @@ export const users = async (req : Request, res : Response, next : NextFunction) 
 
         if(endIndex < await User.find().countDocuments().exec()) result.next = { page : NPage + 1, limit : NLimit }
 
-        if(startIndex > 0) result.previous = { page : NPage - 1, limit : NLimit } 
+        if(startIndex > 0) result.previous = { page : NPage - 1, limit : NLimit }
 
         const users : IUser[] | null = await User.find({isAdmin : false}).limit(NLimit).skip(startIndex)
         .select('-password -isAdmin -token -tokenExpireDate -createdAt -UpdatedAt -__v -profilePic');
@@ -35,7 +35,7 @@ export const users = async (req : Request, res : Response, next : NextFunction) 
         res.status(200).json(users);
 
     } catch (error) {
-        
+
         next(error);
     }
 
@@ -48,52 +48,48 @@ export const deleteUser = async (req : Request, res : Response, next : NextFunct
 
         const user : IUser | null = await User.findById(userToModify);
 
-        await Comment.updateMany<IComment>({
-            $pull : {replies : {userId : userToModify}}
-        });
+        if(!user) return res.status(404).json({error : 'User not found'});
+
+        if(user.isAdmin) return res.status(400).json({error : 'Cannot delete a admin'});
 
         if(user.isSeller) {
 
             const products : IProduct[] | null = await Product.find({user : userToModify});
-            
-            for (const product of products) {
-                
-                await WishList.updateMany<IWishList>({
-                    $pull : {products : {product : product._id}}
-                });
-                
-                await Cart.updateMany<ICart>({
-                    $pull : {products : {product : product._id}}
-                });
 
-                await Inventory.deleteMany({
-                    productId : product._id
-                });
-            };
+            await Promise.all(products.map(async (product) => {
 
-            await Report.updateMany<IReport>({
-                $pull : {reportersId : userToModify}
-            });
-            
-            await Promise.all([Product.deleteMany({user : userToModify}), WishList.deleteMany({user : userToModify}),
-                Cart.deleteMany({user : userToModify}), Order.deleteMany({user : userToModify}), Report.deleteMany({user : userToModify}),
-                Address.deleteMany({user : userToModify}), User.deleteOne({_id : userToModify}), Comment.deleteMany({senderId : userToModify})
-            ]);
+                await WishList.updateMany({}, {$pull : {products : {product : product._id}}});
+                await Cart.updateMany({}, {$pull : {products : {product : product._id}}});
+                await Inventory.deleteMany({productId : product._id});
+            }));
+
+            await Comment.updateMany({$pull : {replies : {userId : userToModify}}});
+
+            await Promise.all([Report.updateMany({}, {$pull : {reportersId : userToModify}}), Product.deleteMany({user : userToModify}),
+                User.deleteOne({_id : userToModify}), Address.deleteOne({user : userToModify}), Comment.deleteMany({senderId : userToModify}),
+                    WishList.deleteOne({user : userToModify}), Cart.deleteOne({user : userToModify})]);
         }
 
-        await Promise.all([WishList.deleteMany({user : userToModify}), Cart.deleteMany({user : userToModify}), 
-            Report.deleteMany({user : userToModify}), Address.deleteMany({user : userToModify}), User.deleteOne({_id : userToModify}),
-            Comment.deleteMany({senderId : userToModify})
+        await Comment.updateMany({$pull : {replies : {userId : userToModify}}});
+
+        await Promise.all([
+            WishList.deleteMany({user: userToModify}), 
+            Cart.deleteMany({ user: userToModify}), 
+            Report.updateMany({}, {$pull : {reportersId : userToModify}}), 
+            Address.deleteMany({user: userToModify}), 
+            Comment.deleteMany({senderId: userToModify}), 
+            User.deleteOne({_id: userToModify})
         ]);
 
         res.status(200).json({message : 'User has been deleted'});
 
     } catch (error) {
-
+        
         next(error);
     }
 
 }
+
 
 export const banUser = async (req : Request, res : Response, next : NextFunction) => {
 
@@ -111,7 +107,7 @@ export const banUser = async (req : Request, res : Response, next : NextFunction
         res.status(200).json({message : 'User has been baned'});
 
     } catch (error) {
-        
+
         next(error);
     }
 
@@ -128,7 +124,7 @@ export const count = async (req : Request, res : Response, next : NextFunction) 
         res.status(200).json({users, products, reports, orders});
 
     } catch (error) {
-        
+
         next(error);
     }
